move solver.Integral to topology.Integral
